feat(replayq): Do not purge all segment files in offload mode reopen
Before this change, in offload mode:
the assumption was that when reading from & writing to the same segment
the in-memory (mem only, no disk copy) segment is lost in case of
restart, and hence data loss is unavoidable, which in turn lead to
the brutal pruge design.

The argument which lead to this change is:
Lossing a few messages in the reading segment is better than purging
all, after all, replayq dropps oldest message even when all on disk
in case of reaching total size limit.